requirements.txt
yt-dlp
streamlink
PyYAML
